Made the reversed hand swap feature more compliant with standards.
