+++ Content.Client/Options/UI/Tabs/KeyRebindTab.xaml.cs
@@ -190,7 +190,7 @@ namespace Content.Client.Options.UI.Tabs
             AddButton(ContentKeyFunctions.Drop);
             AddButton(ContentKeyFunctions.ExamineEntity);
             AddButton(ContentKeyFunctions.SwapHands);
-            AddButton(ContentKeyFunctions.SwapHandsReversed);
+            AddButton(ContentKeyFunctions.SwapHandsReversed); // DeltaV - Swap Hands Reversed
             AddButton(ContentKeyFunctions.MoveStoredItem);
             AddButton(ContentKeyFunctions.RotateStoredItem);
             AddButton(ContentKeyFunctions.SaveItemLocation);
