+++ Content.Shared/Input/ContentKeyFunctions.cs
@@ -35,7 +35,7 @@ namespace Content.Shared.Input
         public static readonly BoundKeyFunction OpenBelt = "OpenBelt";
         public static readonly BoundKeyFunction OpenAHelp = "OpenAHelp";
         public static readonly BoundKeyFunction SwapHands = "SwapHands";
-        public static readonly BoundKeyFunction SwapHandsReversed = "SwapHandsReversed";
+        public static readonly BoundKeyFunction SwapHandsReversed = "SwapHandsReversed"; // DeltaV - Swap Hands Reversed
         public static readonly BoundKeyFunction MoveStoredItem = "MoveStoredItem";
         public static readonly BoundKeyFunction RotateStoredItem = "RotateStoredItem";
         public static readonly BoundKeyFunction SaveItemLocation = "SaveItemLocation";
