+++ Resources/Locale/en-US/_DV/escape-menu/options-menu.ftl
@@ -2,4 +2,4 @@
 ui-options-general-forknotice = Note: These settings are fork-specific and might not apply on other servers.
 
 ui-options-no-filters = Disable species vision filters
-ui-options-function-swap-hands-reversed = Swap Hands Reverse
+ui-options-function-swap-hands-reversed = Swap hands (reversed)
