+++ Resources/keybinds.yml
@@ -165,8 +165,6 @@ binds:
 - function: SwapHands
   type: State
   key: X
-- function: SwapHandsReversed
-  type: State
 - function: MoveStoredItem
   type: State
   key: MouseLeft
